ref(Generative.vue): resuse useGenerativePreview and removed code duplication
