+++ composables/drop/useGenerativeDropMint.ts
@@ -1,5 +1,4 @@
 import { DoResult } from '@/services/fxart'
-import { pinFileToIPFS } from '@/services/nftStorage'
 import useGenerativeIframeData from '@/composables/drop/useGenerativeIframeData'
 
 export type DropMintedNft = DoResult & {
@@ -30,8 +29,6 @@ export default ({
   defaultMax,
   mintedDropCount,
 }: GenerativeDropMintParams) => {
-  const { toast } = useToast()
-  const { $i18n } = useNuxtApp()
   const { imageDataPayload } = useGenerativeIframeData()
   const { listNftByNftWithMetadata, openListingCartModal } =
     useListingCartModal()
@@ -67,27 +64,6 @@ export default ({
 
   const canListMintedNft = computed(() => Boolean(mintedNftWithMetadata.value))
 
-  const tryCapture = async () => {
-    try {
-      const imgFile = await getCaptureImageFile()
-      const imageHash = await pinFileToIPFS(imgFile)
-      return imageHash
-    } catch (error) {
-      toast($i18n.t('drops.capture'))
-      throw error
-    }
-  }
-
-  const getCaptureImageFile = async () => {
-    const selectedImageHash = selectedImage.value.split('?hash=')[1]
-    const isTheSameImage = selectedImageHash === imageDataPayload.value?.hash
-    if (!imageDataPayload.value?.image || !isTheSameImage) {
-      throw new Error('Failed to load image, please try again later')
-    }
-    const res = (await fetch(imageDataPayload.value.image)) as any
-    return new File([res], 'image.png', { type: 'image/png' })
-  }
-
   const subscribeToMintedNft = (id: string, onReady: (data) => void) => {
     useSubscriptionGraphql({
       query: `nftEntityById(id: "${id}") {
@@ -117,7 +93,7 @@ export default ({
     canListMintedNft,
     nftCount,
     listMintedNft,
-    tryCapture,
     subscribeToMintedNft,
+    imageDataPayload,
   }
 }
